requirements.txt
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
decouple==0.0.7
Django==5.1.7
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
urllib3==2.3.0
whitenoise==6.9.0
psycopg2-binary
dj-database-url